+++ Shapes/Rectangle.h
@@ -11,8 +11,8 @@ class Rectangle : public Shape{
 public:
     explicit Rectangle(const double & width, const double & height);
     void intoPS() override;
-    int getWidth() override;
-    int getHeight() override;
+    double getWidth() override;
+    double getHeight() override;
 private:
     double _width;
     double _height;
